pyproject.toml
[project]
name = "micaexp"
version = "0.1.0"
description = ""
authors = [
    {name = "lucash2304", email = "lucash2304@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "dash>=2.17.1,<3.0.0",
    "dash-bootstrap-components>=1.6.0,<2.0.0",
    "plotly>=5.22.0,<6.0.0",
    "pandas>=2.2.2,<3.0.0",
    "numpy>=2.0.0,<3.0.0",
    "scipy>=1.14.0,<2.0.0",
    "matplotlib>=3.9.1,<4.0.0",
    "geopandas>=1.0.1,<2.0.0",
    "shapely>=2.0.5,<3.0.0",
    "rasterio>=1.3.10,<2.0.0",
    "pyproj>=3.6.1,<4.0.0",
    "requests>=2.32.3,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "notebook>=7.2.1,<8.0.0",
    "ipython>=8.26.0,<9.0.0",
    "ipywidgets>=8.1.3,<9.0.0",
    "qtconsole>=5.5.2,<6.0.0",
    "nbconvert>=7.16.4,<8.0.0",
    "nbformat>=5.10.4,<6.0.0",
    "jinja2>=3.1.4,<4.0.0",
    "Flask>=3.0.3,<4.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "api_calls" }
]
